refactor: split overlay view model and timeline helpers
